move CircuitView to MainFragment
