+++ .idea/dictionaries/thssuperiorstanislav.xml
@@ -2,6 +2,7 @@
   <dictionary name="thssuperiorstanislav">
     <words>
       <w>calay</w>
+      <w>thesuperiorstanislav</w>
     </words>
   </dictionary>
 </component>
